move common code to base classes
